requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gevent==24.11.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
setuptools==75.8.0
starkbank-ecdsa==2.2.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
